+++ CHANGELOG.md
@@ -2,8 +2,9 @@ Change Log
 =====
 
 # 0.2.1
-- Require item class name to be provided via abstract method `getItemClassName()` instead of via the collection constructor.
-- Removed support for `ArrayAccess`, and replicate functionality via more meaningful methods such as `Collection::has()` and `Collection::get()`.
+- BC BREAK - Require item class name to be provided via abstract method `getItemClassName()` instead of via the collection constructor.
+- BC BREAK - Removed support for `ArrayAccess`, and replicate functionality via more meaningful methods such as `Collection::has()` and `Collection::get()`.
+- BC BREAK - `Collection::toArray()` renamed to `Collection::all()`.
 - Added the factory methods `Collection::from()` and `Collection::empty()`.
 - Added helper methods `Collection::isEmpty()` and `Collection::isNotEmpty()`.
 - Added `Collection::reverse()` helper to reverse the order of the collection.